requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt --strip-extras requirements.in
#
blinker==1.9.0
    # via flask
charset-normalizer==3.4.7
    # via reportlab
click==8.3.3
    # via flask
deprecated==1.3.1
    # via limits
flask==3.1.3
    # via
    #   -r requirements.in
    #   flask-limiter
flask-limiter==4.1.1
    # via -r requirements.in
gunicorn==25.3.0
    # via -r requirements.in
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
limits==5.8.0
    # via flask-limiter
markupsafe==3.0.3
    # via
    #   flask
    #   jinja2
    #   werkzeug
numpy==2.4.4
    # via scipy
ordered-set==4.1.0
    # via flask-limiter
packaging==26.1
    # via
    #   gunicorn
    #   limits
pillow==12.2.0
    # via
    #   -r requirements.in
    #   reportlab
python-dateutil==2.9.0.post0
    # via
    #   -r requirements.in
    #   rcpchgrowth
rcpchgrowth==4.4.1
    # via -r requirements.in
reportlab==4.4.10
    # via -r requirements.in
scipy==1.17.1
    # via rcpchgrowth
six==1.17.0
    # via python-dateutil
typing-extensions==4.15.0
    # via
    #   flask-limiter
    #   limits
werkzeug==3.1.8
    # via flask
wrapt==2.1.2
    # via deprecated
